+++ index.html
@@ -126,7 +126,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js"
         integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy"
         crossorigin="anonymous"></script>
-    <script src="assets/js/script.js" defer></script>
+    <script type="module" src="assets/js/script.js"></script>
 
     <!-- Service Worker registration handled in assets/js/script.min.js -->
 </body>
