+++ package.json
@@ -16,7 +16,7 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "commonjs",
+  "type": "module",
   "devDependencies": {
     "esbuild": "^0.25.8",
     "jsdom": "^24.1.3",
