+++ test/ensureDiscountToggle.test.js
@@ -1,6 +1,6 @@
-const test = require('node:test');
-const assert = require('node:assert');
-const { JSDOM } = require('jsdom');
+import test from 'node:test';
+import assert from 'node:assert';
+import { JSDOM } from 'jsdom';
 
 test('ensureDiscountToggle inserts a single toggle', () => {
   const dom = new JSDOM(`<!DOCTYPE html><section aria-label="Opciones de filtrado"><div class="row"></div></section>`);
