+++ test/updateProductDisplay.test.js
@@ -1,6 +1,6 @@
-const test = require('node:test');
-const assert = require('node:assert');
-const { JSDOM } = require('jsdom');
+import test from 'node:test';
+import assert from 'node:assert';
+import { JSDOM } from 'jsdom';
 
 // Setup JSDOM environment
 const dom = new JSDOM(`<!DOCTYPE html><div id="product-container"></div>
@@ -106,4 +106,3 @@ test('updateProductDisplay', async (t) => {
     assert.deepStrictEqual(getDisplayedNames(), ['Banana']);
   });
 });
-
